refactor RecipeForm
